feat: 🎸 update setup.zsh, ssh/config

+++ setup/setup.zsh
@@ -10,10 +10,22 @@ fi
 cd ${HOME}/.dotfiles
 
 for name in *; do
-  if [[ ${name} != 'setup' ]] && [[ ${name} != 'README.md' ]]; then
+  if [[ ${name} != 'setup' ]] && [[ ${name} != 'README.md' ]] && [[ ${name} != 'ssh' ]]; then
     if [[ -L ${HOME}/.${name} ]]; then
       unlink ${HOME}/.${name}
     fi
     ln -sfv ${PWD}/${name}  ${HOME}/.${name}
   fi
 done
+
+if [[ ! -d ${HOME}/.ssh ]]; then
+  mkdir ${HOME}/.ssh
+fi
+cd ssh
+for name in *; do
+  if [[ -L ${HOME}/.ssh/$name ]]; then
+    unlink ${HOME}/.ssh/$name
+  fi
+  ln -sfv ${PWD}/${name} ${HOME}/.ssh/${name}
+done
+cd ..
